AI_travel/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
azure-cognitiveservices-speech
pydantic

# LiveKit Agent SDK and plugins
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-api>=0.5.0
aiohttp>=3.9.0
coloredlogs>=15.0.1
